Refactor (moving classes, rename packages)
