+++ app/src/main/java/skarlat/dev/ecoproject/activity/AuthViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
 import skarlat.dev.ecoproject.EcoTipsApp
-import skarlat.dev.ecoproject.authmanager.AuthErrorHandler
-import skarlat.dev.ecoproject.authmanager.DefaultErrorHandler
+import skarlat.dev.ecoproject.authentication.AuthErrorHandler
+import skarlat.dev.ecoproject.authentication.DefaultErrorHandler
 import skarlat.dev.ecoproject.fragment.BaseViewModel
 
 class AuthViewModel : BaseViewModel(), AuthErrorHandler by DefaultErrorHandler() {
