+++ app/src/main/java/skarlat/dev/ecoproject/authentication/AuthError.kt
@@ -1,4 +1,4 @@
-package skarlat.dev.ecoproject.authmanager
+package skarlat.dev.ecoproject.authentication
 
 sealed class AuthError : Throwable(){
     class Default : AuthError()
