+++ app/src/main/java/skarlat/dev/ecoproject/authentication/Authenticator.kt
@@ -1,7 +1,7 @@
-package skarlat.dev.ecoproject.network
+package skarlat.dev.ecoproject.authentication
 
 import android.os.Bundle
-import skarlat.dev.ecoproject.authmanager.strategy.AuthStrategy
+import skarlat.dev.ecoproject.authentication.strategy.AuthStrategy
 
 abstract class Authenticator {
     var authStrategy: AuthStrategy? = null
