+++ app/src/main/java/skarlat/dev/ecoproject/authentication/strategy/GoogleSignInStrategy.kt
@@ -1,4 +1,4 @@
-package skarlat.dev.ecoproject.authmanager.strategy
+package skarlat.dev.ecoproject.authentication.strategy
 
 import android.content.Context
 import android.content.Intent
@@ -9,27 +9,12 @@ import androidx.lifecycle.lifecycleScope
 import com.google.android.gms.auth.api.signin.GoogleSignIn
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions
-import com.google.firebase.auth.FirebaseAuth
 import kotlinx.coroutines.launch
-import skarlat.dev.ecoproject.Const
 import skarlat.dev.ecoproject.EcoTipsApp
 import skarlat.dev.ecoproject.User
-import skarlat.dev.ecoproject.authmanager.AuthError
-import skarlat.dev.ecoproject.authmanager.AuthErrorHandler
-import skarlat.dev.ecoproject.network.AppAuthenticator
-import skarlat.dev.ecoproject.network.user
+import skarlat.dev.ecoproject.authentication.AppAuthenticator
+import skarlat.dev.ecoproject.authentication.AuthError
 import timber.log.Timber
-
-sealed class AuthStrategy {
-    protected val appCache = EcoTipsApp.appComponent.getAppCache()
-    var errorHandler: AuthErrorHandler? = null
-    abstract val authType: AppAuthenticator.AuthMethod
-    abstract fun authenticate(context: Bundle)
-    open fun logout() {
-        appCache.clear()
-    }
-}
-
 
 class GoogleSignInStrategy(activity: AppCompatActivity) : AuthStrategy() {
 
@@ -81,38 +66,4 @@ class GoogleSignInStrategy(activity: AppCompatActivity) : AuthStrategy() {
 
     private val GoogleSignInAccount.user
         get() = User(this.displayName?.substringBefore(" ").orEmpty(), this.email.orEmpty())
-}
-
-class LoginPasswordStrategy : AuthStrategy() {
-    private val fireBaseAuth = FirebaseAuth.getInstance()
-    override val authType: AppAuthenticator.AuthMethod = AppAuthenticator.AuthMethod.PassLogin
-
-    override fun authenticate(context: Bundle) {
-        fireBaseAuth.signInWithEmailAndPassword(context.email, context.password)
-                .addOnFailureListener {
-                    errorHandler?.onError(it)
-                    Timber.e(it, "authLoginPass is fail")
-                }
-                .addOnSuccessListener {
-                    processAuthSuccess()
-                }
-    }
-
-    private fun processAuthSuccess() {
-        fireBaseAuth.currentUser?.let { appCache.setUser(it.user) }
-    }
-
-    override fun logout() {
-        super.logout()
-        fireBaseAuth.signOut()
-    }
-
-
-    private val Bundle.email: String
-        get() = this[Const.AUTH_LOGIN] as? String ?: throw invalidBundleException
-
-    private val Bundle.password: String
-        get() = this[Const.AUTH_PASS] as? String ?: throw invalidBundleException
-
-    private val invalidBundleException = Exception("Invalid bundle for auth")
 }
