+++ app/src/main/java/skarlat/dev/ecoproject/di/AppComponent.kt
@@ -1,8 +1,7 @@
 package skarlat.dev.ecoproject.di
 
-import android.content.Context
 import dagger.Component
-import skarlat.dev.ecoproject.authmanager.AuthManager
+import skarlat.dev.ecoproject.authentication.AuthManager
 import skarlat.dev.ecoproject.core.AppCache
 import javax.inject.Singleton
