+++ app/src/main/java/skarlat/dev/ecoproject/di/AuthModule.kt
@@ -2,9 +2,9 @@ package skarlat.dev.ecoproject.di
 
 import dagger.Module
 import dagger.Provides
-import skarlat.dev.ecoproject.authmanager.AuthManager
+import skarlat.dev.ecoproject.authentication.AppAuthenticator
+import skarlat.dev.ecoproject.authentication.AuthManager
 import skarlat.dev.ecoproject.core.AppCache
-import skarlat.dev.ecoproject.network.AppAuthenticator
 
 @Module(includes = [AppCacheModule::class])
 class AuthModule {
